fix BucketUtil uploadObject function

+++ src/main/java/org/helper_utility/BucketUtil.java
@@ -19,8 +19,12 @@ public class BucketUtil {
 
     public void createBucket(String bucketName) throws Exception {
         Storage storage = StorageOptions.getDefaultInstance().getService();
-        Bucket bucket = storage.create(BucketInfo.of(bucketName));
-        logger.info("Bucket: {} created.", bucket.getName());
+        if (storage.get(bucketName).exists()) {
+            logger.info("Bucket already exists: " + bucketName);
+        } else {
+            Bucket bucket = storage.create(BucketInfo.of(bucketName));
+            logger.info("Bucket: {} created.", bucket.getName());
+        }
     }
 
     public ArrayList<String> listBuckets() throws Exception {
@@ -46,28 +50,28 @@ public class BucketUtil {
         return blobList;
     }
 
-    public void uploadObject(String bucketName, String objectName, String filePath) throws Exception {
+    public void uploadObject(
+            String bucketName, String objectName, String filePath) throws Exception {
         Storage storage = StorageOptions.newBuilder().setProjectId(projectId).build().getService();
         BlobId blobId = BlobId.of(bucketName, objectName);
         BlobInfo blobInfo = BlobInfo.newBuilder(blobId).build();
 
         Storage.BlobWriteOption precondition;
         if (storage.get(bucketName, objectName) == null) {
-            // For a target object that doesn't exist, set the DoesNotExist precondition.
+            // For a target object that does not yet exist, set the DoesNotExist precondition.
             // This will cause the request to fail if the object is created before the request runs.
             precondition = Storage.BlobWriteOption.doesNotExist();
         } else {
             // If the destination already exists in your bucket, instead set a generation-match
-            // precondition. This will cause the request to fail if the existing object's
-            // generation changes before the request runs.
+            // precondition. This will cause the request to fail if the existing object's generation
+            // changes before the request runs.
             precondition =
                     Storage.BlobWriteOption.generationMatch(
-                            storage.get(bucketName, objectName).getGeneration()
-                    );
-            storage.createFrom(blobInfo, Paths.get(filePath), precondition);
-
-            logger.info("File {} uploaded to bucket {} as {}", filePath, bucketName, objectName);
+                            storage.get(bucketName, objectName).getGeneration());
         }
+        storage.createFrom(blobInfo, Paths.get(filePath), precondition);
+
+        logger.info("File {} uploaded to bucket {} as {}", filePath, bucketName, objectName);
     }
 
     public boolean checkIfBucketExists(String bucketName) throws Exception {
@@ -82,9 +86,12 @@ public class BucketUtil {
 
     public void deleteBucket(String bucketName) {
         Storage storage = StorageOptions.newBuilder().setProjectId(projectId).build().getService();
-        Bucket bucket = storage.get(bucketName);
-        bucket.delete();
-
-        logger.info("Bucket: {} deleted.", bucket.getName());
+        if (!storage.get(bucketName).exists()) {
+            logger.info("Bucket does not exist: {}", bucketName);
+        } else {
+            Bucket bucket = storage.get(bucketName);
+            bucket.delete();
+            logger.info("Bucket: {} deleted.", bucket.getName());
+        }
     }
 }
